Renamed model directory to schemas, add .env file, add dotenv, add mongoose connection to local mongodb instance, and update inventory/product POST route schema create logic
