+++ .env
@@ -1,1 +1,1 @@
-MONGO_DB_LOCATION=
+MONGO_DB_LOCATION=mongodb://localhost:27017/inventory
