+++ src/app.js
@@ -3,12 +3,25 @@ const express = require('express');
 const app = express();
 const port = 3000;
 const mongoose = require('mongoose')
-const config = require('dotenv')
+//this gets values from .env file.
+// It uses the value MONGO_DB_LOCATION as the connection string
+require('dotenv').config()
 
-app.get('/',(req,res) => {
+const connect = mongoose.connect(
+    process.env.MONGO_DB_LOCATION, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+});
+
+app.get('/', (req, res) => {
     res.send("Welcome to Abegail\'s Inventory Tracking web application!")
-  })
+})
 
-app.use('/inventory',inventoryRoutes);
-app.listen(port, ()=> console.log(`Listening on port ${port}`));
+app.use('/inventory', inventoryRoutes);
+
+connect.then(
+    () => console.log('Connected :))'),
+    (err) => console.log('to err is human but something happened while trying to connect to mongo :( ...... ' + err)
+);
+app.listen(port, () => console.log(`Listening on port ${port}`));
